backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
google-adk>=1.0,<2
google-genai>=1.0,<2
google-cloud-firestore>=2.19,<3
google-cloud-storage>=2.18,<3
pydantic>=2.10,<3
python-multipart>=0.0.18
python-dotenv>=1.0,<2
firebase-admin>=6.0,<7
